pyproject.toml
[tool.poetry]
name = "studbotty"
version = "0.1.0"
description = "A modular, study-focused AI agent CLI"
authors = ["StudBotty Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.7.0"
ollama = "^0.1.6"
sympy = "^1.12"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
ddgs = "^0.1.3"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
